+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # BSCAMPP v1.0.1b
 1. Removed redundant codes and fixed missing variables.
 2. Added badges for PyPI installation and current Python Build, etc.
+3. Added a single Pytest for dry-running BSCAMPP under `tests/`.
 
 # BSCAMPP v1.0.1a
 1. Completed features with both `epa-ng` and `pplacer` support.